requirements.txt
flask
aiohttp
discord
random
string
pyyaml
asyncio
time
os
lock